Move the prior calculation into the prior module
