+++ src/engine/mod.rs
@@ -22,7 +22,6 @@
 
 pub use self::controller::EngineController;
 pub use self::node::Node;
-pub use self::prior::Prior;
 use board::Board;
 use board::Color;
 use board::Move;
