+++ src/engine/prior/mod.rs
@@ -29,6 +29,7 @@ use std::sync::Arc;
 
 pub struct Prior {
     config: Arc<Config>,
+    m: Move,
     plays: usize,
     wins: usize,
 }
@@ -38,6 +39,7 @@ impl Prior {
     pub fn new(board: &Board, m: &Move, matcher: &Arc<Matcher>, config: Arc<Config>) -> Prior {
         let mut prior = Prior {
             config: config,
+            m: *m,
             plays: 0,
             wins: 0,
         };
@@ -106,4 +108,34 @@ impl Prior {
     fn record_negative_prior(&mut self, prior: usize) {
         self.record_priors(prior, 0);
     }
+}
+
+pub fn calculate(moves: Vec<Move>, board: &Board, matcher: &Arc<Matcher>, config: &Arc<Config>) -> Vec<Prior> {
+    let mut priors: Vec<Prior> = moves.iter()
+        .map(|m| Prior::new(board, m, matcher, config.clone()))
+        .collect();
+    let color = board.next_player().opposite();
+    let in_danger = board.chains().iter()
+        .filter(|chain| {
+            chain.color() == color && chain.coords().len() == 1 && chain.liberties().len() <= 2
+        });
+    for one_stone in in_danger {
+        if let Some(solution) = board.capture_ladder(one_stone) {
+            if let Some(prior) = priors.iter_mut().find(|c| c.m == solution) {
+                prior.record_even_prior(config.priors.capture_one);
+            }
+        }
+    }
+    let in_danger = board.chains().iter()
+        .filter(|chain| {
+            chain.color() == color && chain.coords().len() > 1 && chain.liberties().len() <= 2
+        });
+    for many_stones in in_danger {
+        if let Some(solution) = board.capture_ladder(many_stones) {
+            if let Some(prior) = priors.iter_mut().find(|c| c.m == solution) {
+                prior.record_even_prior(config.priors.capture_many);
+            }
+        }
+    }
+    priors
 }
